Splitting real, imag, angle
