+++ src/awkward/operations/ak_round.py
@@ -5,6 +5,7 @@ from __future__ import annotations
 import awkward as ak
 from awkward._dispatch import high_level_function
 from awkward._layout import HighLevelContext
+from awkward._nplikes.array_module import ArrayLikeT
 from awkward._nplikes.numpy_like import NumpyMetadata
 
 __all__ = ("round",)
